+++ CHANGELOG.md
@@ -15,7 +15,7 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 
 ### Fixed
 
--
+- Fixed bug in next js framework rules
 
 ## [2.4.3] - 2025-12-27